Refactor templateBarsAndLines
