+++ src/util/graph-templates/templateBarsAndLines/generateFlattenData.js
@@ -1,0 +1,9 @@
+export default function(data, scale, key) {
+  return data.map(row => {
+    return {
+      x: this.x(row.date),
+      y: scale(row[key]),
+      label: `${row[key]}`,
+    };
+  });
+}
